backend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
beautifulsoup4==4.12.2
requests==2.32.3
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.1
chromadb==0.6.3
sentence-transformers==2.7.0
groq==0.4.1
python-dotenv==1.0.0